fix: move StMakeEnvelope to its own file to fix scalariform parse error
StMakeEnvelope extends Expression directly (not BinaryExpression) and
uses IndexedSeq[Expression] in withNewChildrenInternal. This unique
structure causes scalariform 0.2.6 to corrupt its AST state, causing
subsequent (g1, g2) => lambdas in the same file to fail to parse.
Moving it to a separate file isolates the issue.
